pytest_infrahub/pyproject.toml
[tool.poetry]
name = "pytest-infrahub"
version = "0.1.0"
description = ""
authors = ["OpsMill <contact@opsmill.com>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://opsmill.io"
repository = "https://opsmill.io"
documentation = "https://opsmill.io"
packages = [
    {include = "pytest_infrahub"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.4"
pyyaml = "^6.0"
pydantic = ">=1.7.4,!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0"

[tool.poetry.plugins."pytest11"]
"pytest-infrahub"="pytest_infrahub.plugin"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
